Implement permalinks for Mercurial.
Cherrypick of 8d4134ba01c9. Pulls vcs parts from omniglot into
dxr.vcs and adds a /rev/ route which loads a particular revision of
a file from VCS.
