+++ dxr/plugins/omniglot.py
@@ -8,6 +8,9 @@ from schema import Optional
 
 import dxr.indexers
 from dxr.plugins import Plugin
+from dxr.vcs import tree_to_repos
+from dxr.utils import DXR_BLUEPRINT
+from flask import url_for
 
 """Omniglot - Speaking all commonly-used version control systems.
 At present, this plugin is still under development, so not all features are
@@ -31,243 +34,6 @@ Todos:
 - check if the mercurial paths are specific to Mozilla's customization or not.
 
 """
-class VCS(object):
-    """A class representing an abstract notion of a version-control system.
-    In general, all path arguments to query methods should be normalized to be
-    relative to the root directory of the VCS.
-    """
-
-    def __init__(self, root):
-        self.root = root
-        self.untracked_files = set()
-
-    def get_root_dir(self):
-        """Return the directory that is at the root of the VCS."""
-        return self.root
-
-    def get_vcs_name(self):
-        """Return a recognizable name for the VCS."""
-        return type(self).__name__
-
-    def invoke_vcs(self, args):
-        """Return the result of invoking said command on the repository, with
-        the current working directory set to the root directory.
-        """
-        return subprocess.check_output(args, cwd=self.get_root_dir())
-
-    def is_tracked(self, path):
-        """Does the repository track this file?"""
-        return path not in self.untracked_files
-
-    def get_rev(self, path):
-        """Return a human-readable revision identifier for the repository."""
-        raise NotImplemented
-
-    def generate_log(self, path):
-        """Return a URL for a page that lists revisions for this file."""
-        raise NotImplemented
-
-    def generate_blame(self, path):
-        """Return a URL for a page that lists source annotations for lines in
-        this file.
-        """
-        raise NotImplemented
-
-    def generate_diff(self, path):
-        """Return a URL for a page that shows the last change made to this file.
-        """
-        raise NotImplemented
-
-    def generate_raw(self, path):
-        """Return a URL for a page that returns a raw copy of this file."""
-        raise NotImplemented
-
-
-class Mercurial(VCS):
-    def __init__(self, root):
-        super(Mercurial, self).__init__(root)
-        # Find the revision
-        self.revision = self.invoke_vcs(['hg', 'id', '-i']).strip()
-        # Sometimes hg id returns + at the end.
-        if self.revision.endswith("+"):
-            self.revision = self.revision[:-1]
-
-        # Make and normalize the upstream URL
-        upstream = urlparse.urlparse(self.invoke_vcs(['hg', 'paths', 'default']).strip())
-        recomb = list(upstream)
-        if upstream.scheme == 'ssh':
-            recomb[0] == 'http'
-        recomb[1] = upstream.hostname # Eliminate any username stuff
-        # check if port is defined and add that to the url
-        if upstream.port:
-            recomb[1] += ":{}".format(upstream.port)
-        recomb[2] = '/' + recomb[2].lstrip('/') # strip all leading '/', add one back
-        if not upstream.path.endswith('/'):
-            recomb[2] += '/' # Make sure we have a '/' on the end
-        recomb[3] = recomb[4] = recomb[5] = '' # Just those three
-        self.upstream = urlparse.urlunparse(recomb)
-
-        # Find all untracked files
-        self.untracked_files = set(line.split()[1] for line in
-            self.invoke_vcs(['hg', 'status', '-u', '-i']).split('\n')[:-1])
-
-        # Determine the last revision on which each file is changed
-        self.previous_revisions = self.find_previous_revisions(root)
-
-    def find_previous_revisions(self, root):
-        """Find the last revision in which each file changed, for diff links.
-
-        Return a mapping {filename: hash of last change commit}
-
-        """
-        from mercurial import hg, ui  # unsupported api
-        repo = hg.repository(ui.ui(), root)
-        last_change = {}
-        for rev in xrange(0, repo['tip'].rev() + 1):
-            ctx = repo[rev]
-            # Go through all filenames changed in this commit:
-            for filename in ctx.files():
-                # str(ctx) gives us the 12-char hash for URLs.
-                last_change[filename] = str(ctx)
-        return last_change
-
-    @classmethod
-    def claim_vcs_source(cls, path, dirs, plugin_config):
-        if '.hg' in dirs:
-            dirs.remove('.hg')
-            return cls(path)
-        return None
-
-    def get_rev(self, path):
-        return self.revision
-
-    def generate_log(self, path):
-        return self.upstream + 'filelog/' + self.revision + '/' + path
-
-    def generate_blame(self, path):
-        return self.upstream + 'annotate/' + self.revision + '/' + path
-
-    def generate_diff(self, path):
-        return self.upstream + 'diff/' + self.previous_revisions[path] + '/' + path
-
-    def generate_raw(self, path):
-        return self.upstream + 'raw-file/' + self.revision + '/' + path
-
-
-class Git(VCS):
-    def __init__(self, root):
-        super(Git, self).__init__(root)
-        self.untracked_files = set(line for line in
-            self.invoke_vcs(['git', 'ls-files', '-o']).split('\n')[:-1])
-        self.revision = self.invoke_vcs(['git', 'rev-parse', 'HEAD'])
-        source_urls = self.invoke_vcs(['git', 'remote', '-v']).split('\n')
-        for src_url in source_urls:
-            name, url, _ = src_url.split()
-            if name == 'origin':
-                self.upstream = self.synth_web_url(url)
-                break
-
-    @classmethod
-    def claim_vcs_source(cls, path, dirs, plugin_config):
-        if '.git' in dirs:
-            dirs.remove('.git')
-            return cls(path)
-        return None
-
-    def get_rev(self, path):
-        return self.revision[:10]
-
-    def generate_log(self, path):
-        return self.upstream + "/commits/" + self.revision + "/" + path
-
-    def generate_blame(self, path):
-        return self.upstream + "/blame/" + self.revision + "/" + path
-
-    def generate_diff(self, path):
-        # I really want to make this anchor on the file in question, but github
-        # doesn't seem to do that nicely
-        return self.upstream + "/commit/" + self.revision
-
-    def generate_raw(self, path):
-        return self.upstream + "/raw/" + self.revision + "/" + path
-
-    def synth_web_url(self, repo):
-        if repo.startswith("git@github.com:"):
-            self._is_github = True
-            return "https://github.com/" + repo[len("git@github.com:"):]
-        elif repo.startswith(("git://github.com/", "https://github.com/")):
-            self._is_github = True
-            if repo.endswith(".git"):
-                repo = repo[:-len(".git")]
-            if repo.startswith("git:"):
-                repo = "https" + repo[len("git"):]
-            return repo
-        raise RuntimeError("Your git remote is not supported yet. Please use a "
-                           "GitHub remote for now, or disable the omniglot "
-                           "plugin.")
-
-
-class Perforce(VCS):
-    def __init__(self, root, upstream):
-        super(Perforce, self).__init__(root)
-        have = self._p4run(['have'])
-        self.have = dict((x['path'][len(root) + 1:], x) for x in have)
-        self.upstream = upstream
-
-    @classmethod
-    def claim_vcs_source(cls, path, dirs, plugin_config):
-        if 'P4CONFIG' not in os.environ:
-            return None
-        if os.path.exists(os.path.join(path, os.environ['P4CONFIG'])):
-            return cls(path, plugin_config.p4web_url)
-        return None
-
-    def _p4run(self, args):
-        ret = []
-        env = os.environ
-        env["PWD"] = self.root
-        proc = subprocess.Popen(['p4', '-G'] + args,
-            stdin=subprocess.PIPE,
-            stdout=subprocess.PIPE,
-            cwd=self.root,
-            env=env)
-        while True:
-            try:
-                x = marshal.load(proc.stdout)
-            except EOFError:
-                break
-            ret.append(x)
-        return ret
-
-    def is_tracked(self, path):
-        return path in self.have
-
-    def get_rev(self, path):
-        info = self.have[path]
-        return '#' + info['haveRev']
-
-    def generate_log(self, path):
-        info = self.have[path]
-        return self.upstream + info['depotFile'] + '?ac=22#' + info['haveRev']
-
-    def generate_blame(self, path):
-        info = self.have[path]
-        return self.upstream + info['depotFile'] + '?ac=193'
-
-    def generate_diff(self, path):
-        info = self.have[path]
-        haveRev = info['haveRev']
-        prevRev = str(int(haveRev) - 1)
-        return (self.upstream + info['depotFile'] + '?ac=19&rev1=' + prevRev +
-                '&rev2=' + haveRev)
-
-    def generate_raw(self, path):
-        info = self.have[path]
-        return self.upstream + info['depotFile'] + '?ac=98&rev1=' + info['haveRev']
-
-
-every_vcs = [Mercurial, Git, Perforce]
-
 
 class TreeToIndex(dxr.indexers.TreeToIndex):
     def pre_build(self):
@@ -276,26 +42,7 @@ class TreeToIndex(dxr.indexers.TreeToIndex):
         keys in ``self.lookup_order``.
 
         """
-        self.source_repositories = {}
-        # Find all of the VCSs in the source directory:
-        for cwd, dirs, files in os.walk(self.tree.source_folder):
-            for vcs in every_vcs:
-                attempt = vcs.claim_vcs_source(cwd, dirs, self.plugin_config)
-                if attempt is not None:
-                    self.source_repositories[attempt.root] = attempt
-
-        # It's possible that the root of the tree is not a VCS by itself, so walk up
-        # the hierarchy until we find a parent folder that is a VCS. If we can't
-        # find any, then no VCSs exist for the top level of this repository.
-        directory = self.tree.source_folder
-        while directory != '/' and directory not in self.source_repositories:
-            directory = os.path.dirname(directory)
-            for vcs in every_vcs:
-                attempt = vcs.claim_vcs_source(directory,
-                                               os.listdir(directory),
-                                               self.plugin_config)
-                if attempt is not None:
-                    self.source_repositories[directory] = attempt
+        self.source_repositories = tree_to_repos(self.tree)
         # Note: we want to make sure that we look up source repositories by deepest
         # directory first.
         self.lookup_order = self.source_repositories.keys()
@@ -324,6 +71,10 @@ class FileToIndex(dxr.indexers.FileToIndex):
             yield 'blame', "Blame", vcs.generate_blame(vcs_relative_path)
             yield 'diff',  "Diff", vcs.generate_diff(vcs_relative_path)
             yield 'raw', "Raw", vcs.generate_raw(vcs_relative_path)
+            yield 'permalink', "Permalink", url_for(DXR_BLUEPRINT + '.permalink',
+                                                    tree=self.tree.name,
+                                                    revision=vcs.get_rev(vcs_relative_path),
+                                                    path=vcs_relative_path)
 
         abs_path = self.absolute_path()
         vcs = self._find_vcs_for_file(abs_path)
